[77] - Move the data management to its own route
